\name{brcadat}
\alias{brcadat}
\docType{data}
\title{RNA-Seq Data of BRCA Patients from TCGA}
\description{
A matrix of RNA-Seq read counts from BRCA patients.
}
\usage{data("brcadat")}
\format{
 A matrix of level 3 RNA-Seq (UNC Illumina HiSeq RNASeqV2) data with 445 breast invasive carcinoma
(BRCA) patients from the Cancer Genome Atlas (TCGA) project on 353 genes with somatic mutations listed in the Catalogue of Somatic Mutations in Cancer (COSMIC). 
The matrix is organized in dimension of \emph{gene} x \emph{sample}.

}
\keyword{datasets}
